requirements.txt
fastapi
uvicorn
tensorflow
pillow
celery
redis
python-multipart
numpy